requirements.txt
asgiref==3.7.2
astroid==3.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
coverage==7.3.3
dill==0.3.7
Django==5.0
django-bootstrap5==23.4
django-filter==23.5
django-recaptcha==4.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
idna==3.6
isort==5.13.2
Markdown==3.5.1
mccabe==0.7.0
Pillow==10.1.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.9.0
urllib3==2.1.0
